Move ComparePopolo class to a separate file
